requirements.txt
aiohttp-retry==2.8.3 ; python_version >= "3.9" and python_version < "3.12"
aiohttp==3.9.1 ; python_version >= "3.9" and python_version < "3.12"
aiosignal==1.3.1 ; python_version >= "3.9" and python_version < "3.12"
anthropic==0.2.10 ; python_version >= "3.9" and python_version < "3.12"
anyio==4.2.0 ; python_version >= "3.9" and python_version < "3.12"
async-timeout==4.0.3 ; python_version >= "3.9" and python_full_version <= "3.11.2"
attrs==23.2.0 ; python_version >= "3.9" and python_version < "3.12"
azure-cognitiveservices-speech==1.34.1 ; python_version >= "3.9" and python_version < "3.12"
certifi==2023.11.17 ; python_version >= "3.9" and python_version < "3.12"
cffi==1.16.0 ; python_version >= "3.9" and python_version < "3.12"
charset-normalizer==3.3.2 ; python_version >= "3.9" and python_version < "3.12"
click==8.1.7 ; python_version >= "3.9" and python_version < "3.12"
colorama==0.4.6 ; python_version >= "3.9" and python_version < "3.12" and platform_system == "Windows"
cryptography==42.0.1 ; python_version >= "3.9" and python_version < "3.12"
dataclasses-json==0.5.14 ; python_version >= "3.9" and python_version < "3.12"
deprecated==1.2.14 ; python_version >= "3.9" and python_version < "3.12"
exceptiongroup==1.2.0 ; python_version >= "3.9" and python_version < "3.11"
fastapi==0.100.1 ; python_version >= "3.9" and python_version < "3.12"
filelock==3.13.1 ; python_version >= "3.9" and python_version < "3.12"
frozenlist==1.4.1 ; python_version >= "3.9" and python_version < "3.12"
fsspec==2023.12.2 ; python_version >= "3.9" and python_version < "3.12"
greenlet==3.0.3 ; python_version >= "3.9" and python_version < "3.12" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32")
h11==0.14.0 ; python_version >= "3.9" and python_version < "3.12"
httpcore==1.0.2 ; python_version >= "3.9" and python_version < "3.12"
httpx==0.26.0 ; python_version >= "3.9" and python_version < "3.12"
huggingface-hub==0.20.3 ; python_version >= "3.9" and python_version < "3.12"
idna==3.6 ; python_version >= "3.9" and python_version < "3.12"
importlib-metadata==6.11.0 ; python_version >= "3.9" and python_version < "3.12"
janus==1.0.0 ; python_version >= "3.9" and python_version < "3.12"
jinja2==3.1.3 ; python_version >= "3.9" and python_version < "3.12"
joblib==1.3.2 ; python_version >= "3.9" and python_version < "3.12"
langchain==0.0.198 ; python_version >= "3.9" and python_version < "3.12"
langchainplus-sdk==0.0.20 ; python_version >= "3.9" and python_version < "3.12"
markupsafe==2.1.4 ; python_version >= "3.9" and python_version < "3.12"
marshmallow==3.20.2 ; python_version >= "3.9" and python_version < "3.12"
miniaudio==1.59 ; python_version >= "3.9" and python_version < "3.12"
multidict==6.0.4 ; python_version >= "3.9" and python_version < "3.12"
mypy-extensions==1.0.0 ; python_version >= "3.9" and python_version < "3.12"
nltk==3.8.1 ; python_version >= "3.9" and python_version < "3.12"
numexpr==2.8.8 ; python_version >= "3.9" and python_version < "3.12"
numpy==1.26.3 ; python_version >= "3.9" and python_version < "3.12"
openai==0.27.10 ; python_version >= "3.9" and python_version < "3.12"
openapi-schema-pydantic==1.2.4 ; python_version >= "3.9" and python_version < "3.12"
opentelemetry-api==1.22.0 ; python_version >= "3.9" and python_version < "3.12"
opentelemetry-sdk==1.22.0 ; python_version >= "3.9" and python_version < "3.12"
opentelemetry-semantic-conventions==0.43b0 ; python_version >= "3.9" and python_version < "3.12"
packaging==23.2 ; python_version >= "3.9" and python_version < "3.12"
pycparser==2.21 ; python_version >= "3.9" and python_version < "3.12"
pydantic==1.10.14 ; python_version >= "3.9" and python_version < "3.12"
pydub==0.25.1 ; python_version >= "3.9" and python_version < "3.12"
pyjwt==2.8.0 ; python_version >= "3.9" and python_version < "3.12"
pyjwt[crypto]==2.8.0 ; python_version >= "3.9" and python_version < "3.12"
pyngrok==6.1.2 ; python_version >= "3.9" and python_version < "3.12"
python-dotenv==1.0.1 ; python_version >= "3.9" and python_version < "3.12"
python-multipart==0.0.6 ; python_version >= "3.9" and python_version < "3.12"
pytz==2023.3.post1 ; python_version >= "3.9" and python_version < "3.12"
pyyaml==6.0.1 ; python_version >= "3.9" and python_version < "3.12"
redis==4.6.0 ; python_version >= "3.9" and python_version < "3.12"
regex==2023.12.25 ; python_version >= "3.9" and python_version < "3.12"
requests==2.31.0 ; python_version >= "3.9" and python_version < "3.12"
scipy==1.12.0 ; python_version >= "3.9" and python_version < "3.12"
six==1.16.0 ; python_version >= "3.9" and python_version < "3.12"
sniffio==1.3.0 ; python_version >= "3.9" and python_version < "3.12"
sounddevice==0.4.6 ; python_version >= "3.9" and python_version < "3.12"
speechrecognition==3.10.1 ; python_version >= "3.9" and python_version < "3.12"
sqlalchemy==2.0.25 ; python_version >= "3.9" and python_version < "3.12"
starlette==0.27.0 ; python_version >= "3.9" and python_version < "3.12"
tenacity==8.2.3 ; python_version >= "3.9" and python_version < "3.12"
tokenizers==0.15.1 ; python_version >= "3.9" and python_version < "3.12"
tqdm==4.66.1 ; python_version >= "3.9" and python_version < "3.12"
twilio==8.12.0 ; python_version >= "3.9" and python_version < "3.12"
typing-extensions==4.9.0 ; python_version >= "3.9" and python_version < "3.12"
typing-inspect==0.9.0 ; python_version >= "3.9" and python_version < "3.12"
urllib3==2.1.0 ; python_version >= "3.9" and python_version < "3.12"
uvicorn==0.21.1 ; python_version >= "3.9" and python_version < "3.12"
vocode==0.1.111 ; python_version >= "3.9" and python_version < "3.12"
vonage-jwt==1.1.0 ; python_version >= "3.9" and python_version < "3.12"
vonage==3.12.0 ; python_version >= "3.9" and python_version < "3.12"
websockets==11.0.3 ; python_version >= "3.9" and python_version < "3.12"
wrapt==1.16.0 ; python_version >= "3.9" and python_version < "3.12"
yarl==1.9.4 ; python_version >= "3.9" and python_version < "3.12"
zipp==3.17.0 ; python_version >= "3.9" and python_version < "3.12"
